Refactor storage and txn modules
