+++ server/src/engine/fractal/test_utils.rs
@@ -35,8 +35,8 @@ use {
         storage::{
             self,
             safe_interfaces::{FSInterface, NullFS, VirtualFS},
+            v1::GNSTransactionDriverAnyFS,
         },
-        txn::gns::GNSTransactionDriverAnyFS,
     },
     parking_lot::{Mutex, RwLock},
     std::collections::HashMap,
@@ -148,7 +148,7 @@ impl<Fs: FSInterface> GlobalInstanceLike for TestGlobal<Fs> {
             space_name, space_uuid, model_name, model_uuid,
         ))?;
         let driver =
-            storage::v1::data_batch::create(&storage::v1::loader::SEInitState::model_path(
+            storage::v1::create_batch_journal(&storage::v1::loader::SEInitState::model_path(
                 space_name, space_uuid, model_name, model_uuid,
             ))?;
         self.model_drivers.write().insert(
