+++ server/src/engine/storage/v1/mod.rs
@@ -1,5 +1,5 @@
 /*
- * Created on Mon May 15 2023
+ * Created on Sat Feb 10 2024
  *
  * This file is a part of Skytable
  * Skytable (formerly known as TerrabaseDB or Skybase) is a free and open-source
@@ -7,7 +7,7 @@
  * vision to provide flexibility in data modelling without compromising
  * on performance, queryability or scalability.
  *
- * Copyright (c) 2023, Sayan Nandan <ohsayan@outlook.com>
+ * Copyright (c) 2024, Sayan Nandan <nandansayan@outlook.com>
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Affero General Public License as published by
@@ -28,25 +28,12 @@
 //!
 //! Target tags: `0.8.0-beta`, `0.8.0-beta.2`, `0.8.0-beta.3`
 
-// impls
-mod batch_jrnl;
-mod journal;
-pub(in crate::engine) mod loader;
-mod rw;
-pub mod spec;
-pub mod sysdb;
-// hl
+mod impls;
 pub mod inf;
-#[cfg(test)]
-mod tests;
+pub mod loader;
+pub mod raw;
 
-// re-exports
-pub(self) use spec::Header;
-
-pub use {
-    journal::{JournalAdapter, JournalWriter},
-    rw::SDSSFileIO,
+pub use self::{
+    impls::gns::GNSTransactionDriverAnyFS, raw::batch_jrnl::create as create_batch_journal,
+    raw::batch_jrnl::DataBatchPersistDriver,
 };
-pub mod data_batch {
-    pub use super::batch_jrnl::{create, DataBatchPersistDriver};
-}
