+++ server/src/engine/storage/v1/raw/batch_jrnl/restore.rs
@@ -24,11 +24,6 @@
  *
 */
 
-use crate::engine::storage::v1::inf::{
-    obj::cell::{self, StorageCellTypeID},
-    DataSource,
-};
-
 use {
     super::{
         MARKER_ACTUAL_BATCH_EVENT, MARKER_BATCH_CLOSED, MARKER_BATCH_REOPEN, MARKER_END_OF_BATCH,
@@ -44,7 +39,13 @@ use {
         idx::{MTIndex, STIndex, STIndexSeq},
         storage::{
             common::interface::fs_traits::FSInterface,
-            v1::rw::{SDSSFileIO, TrackedReader},
+            v1::{
+                inf::{
+                    obj::cell::{self, StorageCellTypeID},
+                    DataSource,
+                },
+                raw::rw::{SDSSFileIO, TrackedReader},
+            },
         },
     },
     std::{
