+++ server/src/engine/storage/v2/mod.rs
@@ -24,5 +24,4 @@
  *
 */
 
-pub mod journal;
-pub mod spec;
+pub(in crate::engine::storage) mod raw;
